Final pull request

+++ src/main/scala/com/itechart/project/routes/AttachmentRoutes.scala
@@ -1,25 +1,27 @@
 package com.itechart.project.routes
 
-import cats.effect.{Blocker, ContextShift, Sync}
+import cats.effect.Sync
 import cats.implicits._
+import com.itechart.project.domain.user.Role
+import com.itechart.project.dto.auth.LoggedInUser
+import com.itechart.project.routes.access.AccessChecker.isResourceAvailable
+import com.itechart.project.routes.response.MarshalResponse.marshalResponse
 import com.itechart.project.services.AttachmentService
 import com.itechart.project.services.error.AttachmentErrors.AttachmentFileError
-import com.itechart.project.services.error.AttachmentErrors.AttachmentFileError.{
-  AttachmentNotFound,
-  InvalidItemAttachment
-}
-import io.chrisdavenport.log4cats.Logger
-import org.http4s.dsl.Http4sDsl
+import com.itechart.project.services.error.AttachmentErrors.AttachmentFileError._
+import fs2.io.file.Files
 import org.http4s.circe.CirceEntityCodec.circeEntityEncoder
-import org.http4s.{EntityEncoder, HttpRoutes, Response, StaticFile}
+import org.http4s.circe.JsonDecoder
+import org.http4s.dsl.Http4sDsl
+import org.http4s.{AuthedRoutes, HttpRoutes, Response, StaticFile}
+import org.typelevel.log4cats.Logger
 
 import scala.util.Try
 
 object AttachmentRoutes {
 
-  def routes[F[_]: Sync: ContextShift: Logger](
-    attachmentService: AttachmentService[F],
-    blocker:           Blocker
+  def routes[F[_]: Sync: Logger: Files](
+    attachmentService: AttachmentService[F]
   ): HttpRoutes[F] = {
     val dsl = new Http4sDsl[F] {}
     import dsl._
@@ -28,48 +30,50 @@ object AttachmentRoutes {
       for {
         file <- attachmentService.findFileById(id)
         response <- file match {
-          case Right(value) => StaticFile.fromFile(value, blocker, Some(request)).getOrElseF(NotFound())
+          case Right(value) => StaticFile.fromFile(value, Some(request)).getOrElseF(NotFound())
           case Left(value)  => NotFound(value.message)
         }
       } yield response
-    }
-
-    def deleteCategory(): HttpRoutes[F] = HttpRoutes.of[F] { case DELETE -> Root / "attachments" / LongVar(id) =>
-      val res = for {
-        deleted <- attachmentService.deleteFile(id)
-      } yield deleted
-
-      marshalResponse(res)
     }
 
     object LongVar {
       def unapply(value: String): Option[Long] = Try(value.toLong).toOption
     }
 
-    def attachmentErrorToHttpResponse(error: AttachmentFileError): F[Response[F]] = {
-      error match {
-        case e: AttachmentNotFound    => NotFound(e.message)
-        case e: InvalidItemAttachment => Conflict(e.message)
+    getAttachment
+  }
 
-        case e => BadRequest(e.message)
-      }
+  def securedRoutes[F[_]: Sync: Logger: JsonDecoder](
+    attachmentService: AttachmentService[F]
+  ): AuthedRoutes[LoggedInUser, F] = {
+    val dsl = new Http4sDsl[F] {}
+    import dsl._
+
+    def deleteCategory(): AuthedRoutes[LoggedInUser, F] = AuthedRoutes.of {
+      case DELETE -> Root / "attachments" / LongVar(id) as user =>
+        if (!isResourceAvailable(user.value.role, List(Role.Manager))) Forbidden()
+        else {
+          val res = for {
+            deleted <- attachmentService.deleteFile(id)
+          } yield deleted
+
+          marshalResponse[F, AttachmentFileError, Boolean](res, attachmentErrorToHttpResponse)
+        }
     }
 
-    def marshalResponse[T](
-      result: F[Either[AttachmentFileError, T]]
-    )(
-      implicit E: EntityEncoder[F, T]
-    ): F[Response[F]] =
-      result
-        .flatMap {
-          case Left(error) => attachmentErrorToHttpResponse(error) <* Logger[F].info("ERROR: " + error.message)
-          case Right(dto)  => Ok(dto)
-        }
-        .handleErrorWith { ex =>
-          InternalServerError(ex.getMessage) <* Logger[F].error(ex.getMessage)
-        }
+    deleteCategory()
+  }
 
-    getAttachment <+> deleteCategory()
+  private def attachmentErrorToHttpResponse[F[_]: Sync: Logger](error: AttachmentFileError): F[Response[F]] = {
+    val dsl = new Http4sDsl[F] {}
+    import dsl._
+
+    error match {
+      case e: AttachmentNotFound    => NotFound(e.message)
+      case e: InvalidItemAttachment => Conflict(e.message)
+
+      case e => BadRequest(e.message)
+    }
   }
 
 }
